+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/api/disconnectedOperations/operations.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { DisconnectedOperationsManagementContext as Client } from "../index.js";
+import type { EdgeContext as Client } from "../index.js";
 import type {
   DisconnectedOperation,
   DisconnectedOperationCreateOrUpdate,
@@ -39,9 +39,7 @@ export function _listDeploymentManifestSend(
   context: Client,
   resourceGroupName: string,
   name: string,
-  options: DisconnectedOperationsListDeploymentManifestOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListDeploymentManifestOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Edge/disconnectedOperations/{name}/listDeploymentManifest{?api%2Dversion}",
@@ -49,7 +47,7 @@ export function _listDeploymentManifestSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -57,10 +55,7 @@ export function _listDeploymentManifestSend(
   );
   return context.path(path).post({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -82,9 +77,7 @@ export async function listDeploymentManifest(
   context: Client,
   resourceGroupName: string,
   name: string,
-  options: DisconnectedOperationsListDeploymentManifestOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListDeploymentManifestOptionalParams = { requestOptions: {} },
 ): Promise<DisconnectedOperationDeploymentManifest> {
   const result = await _listDeploymentManifestSend(context, resourceGroupName, name, options);
   return _listDeploymentManifestDeserialize(result);
@@ -92,15 +85,13 @@ export async function listDeploymentManifest(
 
 export function _listBySubscriptionSend(
   context: Client,
-  options: DisconnectedOperationsListBySubscriptionOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListBySubscriptionOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/providers/Microsoft.Edge/disconnectedOperations{?api%2Dversion}",
     {
       subscriptionId: context.subscriptionId,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -108,10 +99,7 @@ export function _listBySubscriptionSend(
   );
   return context.path(path).get({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -131,32 +119,28 @@ export async function _listBySubscriptionDeserialize(
 /** List DisconnectedOperation resources by subscription ID */
 export function listBySubscription(
   context: Client,
-  options: DisconnectedOperationsListBySubscriptionOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListBySubscriptionOptionalParams = { requestOptions: {} },
 ): PagedAsyncIterableIterator<DisconnectedOperation> {
   return buildPagedAsyncIterator(
     context,
     () => _listBySubscriptionSend(context, options),
     _listBySubscriptionDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-03-15" },
   );
 }
 
 export function _listByResourceGroupSend(
   context: Client,
   resourceGroupName: string,
-  options: DisconnectedOperationsListByResourceGroupOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListByResourceGroupOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Edge/disconnectedOperations{?api%2Dversion}",
     {
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -164,10 +148,7 @@ export function _listByResourceGroupSend(
   );
   return context.path(path).get({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -188,16 +169,14 @@ export async function _listByResourceGroupDeserialize(
 export function listByResourceGroup(
   context: Client,
   resourceGroupName: string,
-  options: DisconnectedOperationsListByResourceGroupOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsListByResourceGroupOptionalParams = { requestOptions: {} },
 ): PagedAsyncIterableIterator<DisconnectedOperation> {
   return buildPagedAsyncIterator(
     context,
     () => _listByResourceGroupSend(context, resourceGroupName, options),
     _listByResourceGroupDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-03-15" },
   );
 }
 
@@ -213,7 +192,7 @@ export function _$deleteSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -250,6 +229,7 @@ export function $delete(
     abortSignal: options?.abortSignal,
     getInitialResponse: () => _$deleteSend(context, resourceGroupName, name, options),
     resourceLocationConfig: "location",
+    apiVersion: context.apiVersion ?? "2026-03-15",
   }) as PollerLike<OperationState<void>, void>;
 }
 
@@ -266,7 +246,7 @@ export function _updateSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -275,10 +255,7 @@ export function _updateSend(
   return context.path(path).patch({
     ...operationOptionsToRequestParameters(options),
     contentType: "application/json",
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
     body: disconnectedOperationUpdateSerializer(properties),
   });
 }
@@ -313,9 +290,7 @@ export function _createOrUpdateSend(
   resourceGroupName: string,
   name: string,
   resource: DisconnectedOperationCreateOrUpdate,
-  options: DisconnectedOperationsCreateOrUpdateOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsCreateOrUpdateOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Edge/disconnectedOperations/{name}{?api%2Dversion}",
@@ -323,7 +298,7 @@ export function _createOrUpdateSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -332,10 +307,7 @@ export function _createOrUpdateSend(
   return context.path(path).put({
     ...operationOptionsToRequestParameters(options),
     contentType: "application/json",
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
     body: disconnectedOperationCreateOrUpdateSerializer(resource),
   });
 }
@@ -359,9 +331,7 @@ export function createOrUpdate(
   resourceGroupName: string,
   name: string,
   resource: DisconnectedOperationCreateOrUpdate,
-  options: DisconnectedOperationsCreateOrUpdateOptionalParams = {
-    requestOptions: {},
-  },
+  options: DisconnectedOperationsCreateOrUpdateOptionalParams = { requestOptions: {} },
 ): PollerLike<
   OperationState<DisconnectedOperationCreateOrUpdate>,
   DisconnectedOperationCreateOrUpdate
@@ -372,6 +342,7 @@ export function createOrUpdate(
     getInitialResponse: () =>
       _createOrUpdateSend(context, resourceGroupName, name, resource, options),
     resourceLocationConfig: "azure-async-operation",
+    apiVersion: context.apiVersion ?? "2026-03-15",
   }) as PollerLike<
     OperationState<DisconnectedOperationCreateOrUpdate>,
     DisconnectedOperationCreateOrUpdate
@@ -390,7 +361,7 @@ export function _getSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -398,10 +369,7 @@ export function _getSend(
   );
   return context.path(path).get({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }