+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/api/hardwareSettings/index.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { $delete, createOrUpdate, get, listByParent } from "./operations.js";
+export {
+  HardwareSettingsDeleteOptionalParams,
+  HardwareSettingsCreateOrUpdateOptionalParams,
+  HardwareSettingsGetOptionalParams,
+  HardwareSettingsListByParentOptionalParams,
+} from "./options.js";