+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/api/images/operations.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import type { DisconnectedOperationsManagementContext as Client } from "../index.js";
+import type { EdgeContext as Client } from "../index.js";
 import type { _ImageListResult, Image, ImageDownloadResult } from "../../models/models.js";
 import {
   errorResponseDeserializer,
@@ -34,7 +34,7 @@ export function _listDownloadUriSend(
       resourceGroupName: resourceGroupName,
       name: name,
       imageName: imageName,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -42,10 +42,7 @@ export function _listDownloadUriSend(
   );
   return context.path(path).post({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -88,7 +85,7 @@ export function _getSend(
       resourceGroupName: resourceGroupName,
       name: name,
       imageName: imageName,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
     },
     {
       allowReserved: options?.requestOptions?.skipUrlEncoding,
@@ -96,10 +93,7 @@ export function _getSend(
   );
   return context.path(path).get({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -130,9 +124,7 @@ export function _listByDisconnectedOperationSend(
   context: Client,
   resourceGroupName: string,
   name: string,
-  options: ImagesListByDisconnectedOperationOptionalParams = {
-    requestOptions: {},
-  },
+  options: ImagesListByDisconnectedOperationOptionalParams = { requestOptions: {} },
 ): StreamableMethod {
   const path = expandUrlTemplate(
     "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Edge/disconnectedOperations/{name}/images{?api%2Dversion,%24filter,%24top,%24skip}",
@@ -140,7 +132,7 @@ export function _listByDisconnectedOperationSend(
       subscriptionId: context.subscriptionId,
       resourceGroupName: resourceGroupName,
       name: name,
-      "api%2Dversion": context.apiVersion,
+      "api%2Dversion": context.apiVersion ?? "2026-03-15",
       "%24filter": options?.filter,
       "%24top": options?.top,
       "%24skip": options?.skip,
@@ -151,10 +143,7 @@ export function _listByDisconnectedOperationSend(
   );
   return context.path(path).get({
     ...operationOptionsToRequestParameters(options),
-    headers: {
-      accept: "application/json",
-      ...options.requestOptions?.headers,
-    },
+    headers: { accept: "application/json", ...options.requestOptions?.headers },
   });
 }
 
@@ -176,15 +165,13 @@ export function listByDisconnectedOperation(
   context: Client,
   resourceGroupName: string,
   name: string,
-  options: ImagesListByDisconnectedOperationOptionalParams = {
-    requestOptions: {},
-  },
+  options: ImagesListByDisconnectedOperationOptionalParams = { requestOptions: {} },
 ): PagedAsyncIterableIterator<Image> {
   return buildPagedAsyncIterator(
     context,
     () => _listByDisconnectedOperationSend(context, resourceGroupName, name, options),
     _listByDisconnectedOperationDeserialize,
     ["200"],
-    { itemName: "value", nextLinkName: "nextLink" },
+    { itemName: "value", nextLinkName: "nextLink", apiVersion: context.apiVersion ?? "2026-03-15" },
   );
 }