+++ sdk/disconnectedoperations/arm-disconnectedoperations/src/api/index.ts
@@ -1,8 +1,4 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-export {
-  createDisconnectedOperationsManagement,
-  DisconnectedOperationsManagementContext,
-  DisconnectedOperationsManagementClientOptionalParams,
-} from "./disconnectedOperationsManagementContext.js";
+export { createEdge, EdgeContext, EdgeClientOptionalParams } from "./edgeContext.js";